identity/service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyJWT==2.9.0
bcrypt==4.2.0
pydantic==2.8.2
python-jose[cryptography]==3.3.0
cryptography==41.0.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.2